[update] 舊版資料匯入調整

+++ src/EtheriT.Coker.Web.ConsoleApp/Models/OldDB/ShopInfo.cs
@@ -38,16 +38,16 @@ namespace EtheriT.Coker.Web.ConsoleApp.Models.OldDB
         public string? Picdescribe3 { get; set; }
         public string? Map { get; set; }
         public string? Gov { get; set; }
-        public float? Px { get; set; }
-        public float? Py { get; set; }
+        public double? Px { get; set; }
+        public double? Py { get; set; }
         public string? orgClass { get; set; }
         public string? Class1 { get; set; }
         public string? Class2 { get; set; }
         public string? Class3 { get; set; }
         public string? Level { get; set; }
         public string? Parkinginfo { get; set; }
-        public float? Parkinginfo_px { get; set; }
-        public float? Parkinginfo_py { get; set; }
+        public double? Parkinginfo_px { get; set; }
+        public double? Parkinginfo_py { get; set; }
         public string? Ticketinfo { get; set; }
         public string? Remarks { get; set; }
         public string? Keyword { get; set; }
@@ -65,10 +65,10 @@ namespace EtheriT.Coker.Web.ConsoleApp.Models.OldDB
         public int? TripAdvisorCommentNum { get; set; }
         public string? TripAdvisorUrl { get; set; }
         public int? TripAdvisorRank { get; set; }
-        public float? TripAdvisorComment { get; set; }
+        public double? TripAdvisorComment { get; set; }
         public int? googleCommentNum { get; set; }
         public string? googleUrl { get; set; }
-        public float? googleComment { get; set; }
+        public double? googleComment { get; set; }
         public int? duration { get; set; }
         public string? busInfo { get; set; }
         public string? BeaconMajor { get; set; }
